refactor user nameGenerator codes
